refactor: give input registry APIs their own mod
