+++ crates/zizmor/src/audit/artipacked.rs
@@ -188,7 +188,7 @@ mod tests {
     use crate::{
         github_api::GitHubHost,
         models::{AsDocument, workflow::Workflow},
-        registry::InputKey,
+        registry::input::InputKey,
         state::AuditState,
     };
